Add return_response_on_nxdomain bool to parameters
When using DNSSEC and you get NXDOMAIN, you need to get the NSEC/NSEC3
records to validate the absense of the zone.  For that, you need to get
the answer, which the NXDOMAIN exception does not provide. Setting
return_response_on_nxdomain to True provides a way to get that answer.