test: incorrect test relying on type comparison
This test relies on python2 type comparison mechanism.

A simplier test case is:
``` python
In [1]: class X(object):
    def __gt__(self, other):
        return NotImplemented
   ...:

In [2]: a=X()

In [3]: a>1
Out[3]: True

In [4]: a<1
Out[4]: False
```

Works on python2, not on python3.
Therefore, I'm dropping the testcase

See http://eli.thegreenplace.net/2016/comparing-types-in-python-3/ for
more information about that

Signed-off-by: Arthur Gautier <baloo@gandi.net>