Fixing dns.inet.is_multicast() for Python 3
In Python 3, the dns.ipv4.inet_aton() returns a Bytes object, while in
Python 2 it returns a String. Because of that, with Python 3, the
dns.inet.is_multicast() function fails when processing the object
returned by dns.ipv4.inet_aton().
This commit fixes the behavior by running different block of code for
Python 3.