Improve TCP connect behavior.
Before this change, the _connect() method would start the connection
process, but not wait for it to complete.  This would leave the socket
in an indeterminate state until some other code checked for writability,
and would lose the error code if the connect failed.

This changes _connect() to wait for the connection to complete, and
raises and exception with the appropriate error code if it fails.