Add reverse_query type hinting
Add the type hinting for reverse_query.  `*args` and `**kwargs` datatypes can be any of the data types that are accepted, in theory `Optional[Union[int,str,bool]]` will cover all use cases for any of the arguments passed in (though I have not fully tested this, mypy, pylint, etc. do not complain about this being used in this format)