Remove choose_relativity() from zone.from_xfr()
The comment states that relativize must be consistent between
dns.query.xfr() and dns.zone.from_xfr(), and the code fails if they're
not (if check_origin is True, at least).  This means that the rdata is
already correctly relativized (or not).

This also adds a test of creating zones from xfrs, both relativized and
not.