Implement support for SVCB and HTTPS records
This implements the SVBC and HTTPS records as described in
https://tools.ietf.org/html/draft-ietf-dnsop-svcb-https-00

Given that the dnspython library is often used to test other
implementations, it is useful for it to implement this draft, even
though it's not finalized yet, to help with interop testing.

The wire and presentation formats for SVCB and HTTPS are identical,
so the base RD implementation is split into its own class, so that it
can be reused for all records that have the same format.