Improve TXT records in example zone.
This zone was copied from a zone file in BIND long ago.  BIND contains
both sample zone files and a script to generate sample zone files, but
the committed files appear to have been generated by a broken version of
the script.

This updates the TXT records to what the script would have generated,
which is far more complete.