Fix rdata pickling.
Coverage testing showed that while rdatas would pickle and unpickle
apparently successfully, in fact only the slots from the deepest class
in the inheritance chain would be restored.  So, e.g., a restored A rdata
would have an address attribute but no rdclass or rdtype attributes,
and so things like rdata comparison would break.

This change preserves the whole set of slots, from all ancestors as
well as the object, as a dictionary.