Attempt to refactor per-opcode validation.
Instead of validating rrsets and sections after parsing them, check the
class/type for each record before parsing it.  This is more generic,
because it moves all of the update logic out of the common code.  It's
also more flexible, as it allows the update logic to specify that
meta-records are empty.