Remove the concept from "first" from TSIG.
The sign() and validate() routines took a "first" parameter, which
indicated that this message was the first in a multi-message sequence.
This isn't needed, as it's identical to "not (ctx and multi)".

Remove the parameter from both, as well as the now-unneeded field in the
message object and message.from_wire() parameter.