Make dns.rdata._base64ify(..., 0) work.
In some cases, the caller absolutely doesn't want word breaks.  This
shouldn't be the case for any normal DNS record, but is for records that
don't have well-defined text formats, like TSIG and TKEY.  Allow them to
pass 0 (or None), to indicate that no word breaks should be added.

Previously, passing either 0 or None resulted in an exception, as the
value was used directly as the step in a slice.