The from_text_list() methods for Rdataset and RRset should allow the full set of parameters that rdata's from_text() allows (i.e. origin, relativize, and relativize_to).
These parameters were added though the order is slightly different
from dns.rdata.to_text() as in the the rdata version idna_codec is
after origin, relativize, and relativize_to.  We already had an
idna_codec for the Rdataset and RRset functions, so we added after it
to keep backwards compatibility.