use of uint32 in SOA
If I understand RFC1035 and the clarifications in  RFC 2181, Clarifications to the DNS Standard, section 8 correct,
only the TTL values of ressource records are to be limited to 2^31-1 but REFRESH, RETRY, EXPIRE and MINIMUM in 
the SOA are valid to be 32 bit unsigned integers. As powerdns, knot and dig allow such values and there are nameservers on the internet
serving such SOA records, I propose to allow them here as well.