Fix httpx verify bug.
If both source and verify are passed to dns.query.https() when using
httpx, the verify parameter is ignored.  This is because the code
creates a custom transport for the source address, and httpx only uses the
verify parameter when creating a transport.

The fix is to pass in the verify parameter when we create a transport.