Fix passing source address with httpx.
When dns.query.https() is passed a source address, it needs to create an
adapter for either requests or httpx to use it.  As it doesn't know
which one will actually be used, it creates both.  But then if it
decides to use httpx, it attempts to attach the requests adapter, which
fails.