Fix unintended "wait forever" behavior with zero timeouts [#976].
In a few places we did "if timeout:" or "if expiration:" when we
really meant "if timeout is not None:".  This meant that in the zero
timeout case we fell into the "wait forever" path instead of
immediately timing out.  In the case of UDP queries, we'd be waiting
on recvfrom() and if a packet was lost, then the code would never wake
up.

(cherry picked from commit 0c183f10c78941a4e72046d4dcb2ecf20083b398)