Fix setting source for sync/trio quic queries.
The sync code called connect() before bind(), which meant that any
attempt to specify a source resulted in an exception.  This switches the
order.

The trio code called a nonexistent method in the wrong place, so didn't
work at all.  This fixes the call and puts it in the right place.

The asyncio code worked, so no changes were needed.