Replace asserts with conditions
Running the python interpreter with the optimization mode turned on
(`python -O`, `python -OO`, `set PYTHONOPTIMIZE=1`), will remove
all the assert statements from running code [1][2][3].

Dnspython heavily rely on `assert` to check things at runtime.
That could be an issue, if the flag is enabled. All the dnspython
check would be removed, leading to possible bugs.

This patch propose to replace `assert` statements with condition and
by raising an `AssertionError` if the condition is met.

This patch won't change the logic and the behaviour of the checks
previously made by using `assert`, however, these changes will remain
actives at runtime, even if python optimization is enabled.

This patch do not replace `assert` in unit tests because they are
not an issue.

[1] https://docs.python.org/3/using/cmdline.html?highlight=pythonoptimize#cmdoption-O
[2] https://docs.python.org/3/using/cmdline.html?highlight=pythonoptimize#cmdoption-OO
[3] https://docs.python.org/3/using/cmdline.html?highlight=pythonoptimize#envvar-PYTHONOPTIMIZE