Fix elliptic curve test deprecation warning from cryptography 42.
We were passing a curve class as the curve parameter in
testSignatureECDSAP256SHA256 and testSignatureECDSAP384SHA384,
not an instance of the curve class.  The official API has
always been to pass an instance, but it tolerated passing a class.
Starting with Cryptogrphy 42, passing a class is deprecated.

(cherry picked from commit 8131d0cf4254d9d42f6e23cde3bcc95ff034d341)