Add a copy mode to dns.message.make_response(). (#1131)
Add a copy mode to dns.message.make_response().

If the mode is none, then a default copy mode appropriate for the opcode will
be used.  This is currently always dns.message.CopyMode.QUESTION.

If the mode is dns.message.CopyMode.QUESTION then only the question
section is copied.

If the mode is dns.message.CopyMode.EVERYTHING, then all sections are
copied other than OPT or TSIG records which are created appropriately
if needed instead of being copied.

If the mode is dns.message.CopyMode.NOTHING then no sections are
copied.