More TSIG fixes. (#1207)
* More TSIG fixes.

This attempts to fix the bug that when a message containing a TSIG
record is parsed from wire format, attempting to render it back to wire
format will either regenerate the TSIG (if the TSIG was verified) or
throw an exception (if the TSIG was not verified).  In either case,
the rendered message should contain the TSIG record that was parsed.

* Fix setting tsig_ctx.