Make better APIs for creating sockets and TLS contexts (#1217)
* Make better APIs for creating sockets and TLS contexts for
code that wants to have persistent connections.  [#1176].

This code keeps the ability to use dnspython when the ssl module
doesn't work, but moves the helper code to another module to declutter
and make testing and type checking easier.

We still have to make some type checking compromises, but we are
making fewer than before.